requirements.txt
pandas
influxdb-client
requests
python-dateutil
pytz

# dev
isort
black
